Create an explicit summary object called TestResults.
Doesn't really make the code much simpler, with the exception that it allows
summaries-of-summaries to be calculated more easily. These can replace the
formulae in Excel, but it is mainly done with an eye to eventual HTML output
where creating sums and ratios would be more involved. Best to do it in one
place.